% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_flagged_arguments}
\alias{format_flagged_arguments}
\title{Format Flagged Arguments}
\usage{
format_flagged_arguments(input_args)
}
\arguments{
\item{input_args}{an unnamed vector of arguments of the form "mode=negative"}
}
\value{
a named list
}
\description{
Convert a command-line input / json input a named list.
}
\examples{
format_flagged_arguments(c("MS1tol=10ppm", "MS2tol=20ppm", "mode=negative", "fu='b a r'"))

}
