% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{mzkit_config_file_from_default}
\alias{mzkit_config_file_from_default}
\title{Create mzkit config file from defaults}
\usage{
mzkit_config_file_from_default(
  default_config_path,
  methodId,
  export_file_path,
  rt_alignment_file_override = "",
  spline_ridge_penalty = 20,
  is_alignment = TRUE,
  is_splitting = FALSE,
  is_coelution = TRUE,
  is_search = TRUE,
  is_qc = TRUE,
  is_eda = TRUE,
  reference_sample_override = NULL,
  is_exclude_isotopes = FALSE
)
}
\arguments{
\item{default_config_path}{path to a default config file}

\item{methodId}{mass spectrometry method being run (e.g., M001A)}

\item{export_file_path}{path to new config file}

\item{rt_alignment_file_override}{path to an alignment file or an empty string to ignore}

\item{spline_ridge_penalty}{spline_ridge_penalty Ridge penalty on sample-level splines which model deviations between observed retention times and the consensus rt of a compound.}

\item{is_alignment}{should alignment be performed?}

\item{is_splitting}{should splitting be corrected?}

\item{is_coelution}{should coelutions be found?}

\item{is_search}{should search be performed?}

\item{is_qc}{should QC be performed?}

\item{is_eda}{should EDA be performed?}

\item{reference_sample_override}{NULL to skip or the name of a reference sample}

\item{is_exclude_isotopes}{should isotopologues be excluded?}
}
\value{
writes a json config to \code{export_file_path}
}
\description{
Create mzkit config file from defaults
}
