% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{pipeline_aggregate_split_peaks}
\alias{pipeline_aggregate_split_peaks}
\title{Pipeline Aggregate Split Peaks}
\usage{
pipeline_aggregate_split_peaks(...)
}
\arguments{
\item{...}{all arguments are passed in from mzkit.R using \code{do.call(rwrapper, as.list(formatted_flags))}}
}
\value{
0 invisibly, function called for side effects.
}
\description{
Identifies and merges split peaks caused by the same analyte being picked up in
different peakgroups across samples.
}
