% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{pipeline_alignment}
\alias{pipeline_alignment}
\title{Pipeline Alignment}
\usage{
pipeline_alignment(...)
}
\arguments{
\item{...}{all arguments are passed in from mzkit.R using \code{do.call(rwrapper, as.list(formatted_flags))}}
}
\value{
0 invisibly, function called for side effects.
}
\description{
Align retention time across samples using M2-based alignment
}
