% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{pipeline_coelution_detection}
\alias{pipeline_coelution_detection}
\title{Pipeline Coelution Detection}
\usage{
pipeline_coelution_detection(...)
}
\arguments{
\item{...}{all arguments are passed in from mzkit.R using \code{do.call(rwrapper, as.list(formatted_flags))}}
}
\value{
0 invisibly, function called for side effects.
}
\description{
Summarizes an mzDeltas output to identify coelution relationships between peaks.
Output tables are added to the mzrollDB database.
}
