% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{pipeline_coelution_labeling}
\alias{pipeline_coelution_labeling}
\title{Pipeline Coelution Labeling}
\usage{
pipeline_coelution_labeling(...)
}
\arguments{
\item{...}{all arguments are passed in from mzkit.R using \code{do.call(rwrapper, as.list(formatted_flags))}}
}
\value{
0 invisibly, function called for side effects.
}
\description{
Groups coeluting peakgroups which are likely to be adducts or isotopologues of one another.
}
