% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{pipeline_direct_infusion}
\alias{pipeline_direct_infusion}
\title{Pipeline Direct Infusion}
\usage{
pipeline_direct_infusion(...)
}
\arguments{
\item{...}{all arguments are passed in from mzkit.R using \code{do.call(rwrapper, as.list(formatted_flags))}}
}
\value{
0 invisibly, function called for side effects.
}
\description{
Identify compounds in direct infusion data
}
