% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{pipeline_eda}
\alias{pipeline_eda}
\title{Pipeline EDA}
\usage{
pipeline_eda(...)
}
\arguments{
\item{...}{all arguments are passed in from mzkit.R using \code{do.call(rwrapper, as.list(formatted_flags))}}
}
\value{
0 invisibly, function called for side effects.
}
\description{
Save heatmaps of all compounds and just identified compounds
}
