% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{pipeline_qc}
\alias{pipeline_qc}
\title{Pipeline QC}
\usage{
pipeline_qc(...)
}
\arguments{
\item{...}{all arguments are passed in from mzkit.R using \code{do.call(rwrapper, as.list(formatted_flags))}}
}
\value{
0 invisibly, function called for side effects.
}
\description{
Create a QC directory containing results from internal and external standards.
}
