% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_search.R
\name{pipeline_standard_search}
\alias{pipeline_standard_search}
\title{Pipeline Standard Search}
\usage{
pipeline_standard_search(...)
}
\arguments{
\item{...}{all arguments are passed in from mzkit.R using \code{do.call(rwrapper, as.list(formatted_flags))}}
}
\value{
0 invisibly, function called for side effects.
}
\description{
Identify peakgroups which match known compounds
}
\details{
Using a matching model compare experimental features of peakgroups to candidate compounds which they may match. The mzrollDB file is updated
by annotating peakgroups with their best match and adding all matches
as a separate table. Feasible matches are written out as a .msp file
to support manual validation in MAVEN.
}
