% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{reclass_cmd_args}
\alias{reclass_cmd_args}
\title{Reclass commmand line args}
\usage{
reclass_cmd_args(chr_list, call_fxn, default_args = NULL, other_args = NULL)
}
\arguments{
\item{chr_list}{a named list of specified arguments of class character.}

\item{call_fxn}{an unquoted function which is meant to be called with the processed arguments.}

\item{default_args}{a list containing default values of arguments to be used if a value is not specified in \code{chr_list}.}

\item{other_args}{a named vector containing the classes of other arguments (besides \code{default_args}) whose class should be set.}
}
\value{
a named list of arguments to be supplied to \code{call_fxn}
}
\description{
Using a list of character arguments (generated from parsing class-less command line inputs),
change the class of inputs based on default arguments, and class specifications, then pad the list with the
unspecified defaults.
}
\details{
Currently vector arguments are not supported.
}
\examples{
reclass_cmd_args(chr_list = list(cosine_cutoff = "0.96", print_plots = "TRUE", show_specific_samples = "1", fu = "bar"),
                 call_fxn = clamr::ms2_driven_aligner,
                 default_args = list(group_by_charge = FALSE, maximum_mz_set_size = 200L, cosine_cutoff = 0.95, spline_ridge_penalty = 200, print_plots = TRUE),
                 other_args = c(show_specific_samples = "integer"))

}
