% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{run_quahog}
\alias{run_quahog}
\title{Run Quahog}
\usage{
run_quahog(
  data_folder,
  output_folder = "/tmp/metabolomics_test",
  methodId,
  config_args = list(),
  path_to_mass_spec = "~/mass_spec",
  conda_env,
  conda_path,
  background = FALSE
)
}
\arguments{
\item{data_folder}{directory containing sample mzXML/mzML files.}

\item{output_folder}{directory where output files should be saved.}

\item{methodId}{mass spectrometry method being run (e.g., M001A)}

\item{config_args}{list of config arguments to change from the method-specific defaults}

\item{path_to_mass_spec}{path to mass_spec github repo}

\item{conda_env}{conda environment to use when running quahog}

\item{conda_path}{path to conda directory}

\item{background}{run process in the background?}
}
\value{
a proc object for the pipeline running in the background
}
\description{
Run the quahog pipeline directly from R with a configurable config file.
}
\details{
This is only meant as a convenience function for running mzkit with
an appropriate config during development. The primary methods
for running quahog is via Docker.
}
