% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{sideload_json_config}
\alias{sideload_json_config}
\title{Sideload json config}
\usage{
sideload_json_config(
  rwrapper,
  output_folder,
  config_path = NULL,
  r_scripts_path
)
}
\arguments{
\item{rwrapper}{string corresponding to R module to run}

\item{output_folder}{directory where output files should be saved.}

\item{config_path}{path to a config file (or NULL if config is output_folder/config.json)}

\item{r_scripts_path}{path to mass_spec/lib/R}
}
\value{
a named character vector of rwrapper arguments
}
\description{
Load an quahog configuration file
}
