% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{test_quahog_packages}
\alias{test_quahog_packages}
\title{Test Quahog Packages}
\usage{
test_quahog_packages()
}
\description{
Determine whether the R packages required to run the quahog pipeline are installed.
}
